# v0.6.1: minor cleanups