--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,40 +1,38 @@
 #![forbid(unsafe_code)]
 
-mod backend;
+#[cfg(feature = "__async")]
+pub mod async_reader;
+#[cfg(feature = "http-async")]
+mod backend_http;
+#[cfg(feature = "mmap-async-tokio")]
+mod backend_mmap;
+#[cfg(feature = "__async-s3")]
+mod backend_s3;
+#[cfg(feature = "__async")]
+pub mod cache;
 mod directory;
 mod error;
 mod header;
+#[cfg(feature = "__async")]
 mod tile;
 
-#[cfg(any(
-    feature = "http-async",
-    feature = "mmap-async-tokio",
-    feature = "s3-async-rustls",
-    feature = "s3-async-native"
-))]
-pub mod async_reader;
-
-#[cfg(any(
-    feature = "http-async",
-    feature = "mmap-async-tokio",
-    feature = "s3-async-native",
-    feature = "s3-async-rustls"
-))]
-pub mod cache;
-
 #[cfg(feature = "http-async")]
-pub use backend::HttpBackend;
+pub use backend_http::HttpBackend;
 #[cfg(feature = "mmap-async-tokio")]
-pub use backend::MmapBackend;
-#[cfg(any(feature = "s3-async-rustls", feature = "s3-async-native"))]
-pub use backend::S3Backend;
+pub use backend_mmap::MmapBackend;
+#[cfg(feature = "__async-s3")]
+pub use backend_s3::S3Backend;
 pub use directory::{DirEntry, Directory};
 pub use error::{PmtError, PmtResult};
 pub use header::{Compression, Header, TileType};
+//
+// Re-export crates exposed in our API to simplify dependency management
 #[cfg(feature = "http-async")]
 pub use reqwest;
-#[cfg(any(feature = "s3-async-rustls", feature = "s3-async-native"))]
+#[cfg(feature = "__async-s3")]
 pub use s3;
+#[cfg(feature = "tilejson")]
+pub use tilejson;
 
 #[cfg(test)]
 mod tests {